Move update functionality into remove script
